+++ src/components/background/RoundedBackground.tsx
