+++ src/components/feature/VerticalFeatureRow.tsx
