+++ src/components/template/Logo.tsx
@@ -1,4 +1,4 @@
-import { AppConfig } from '../utils/AppConfig';
+import { AppConfig } from '../../utils/AppConfig';
 
 type ILogoProps = {
   xl?: boolean;
