+++ src/components/template/VerticalFeatures.tsx
